FEAT: Added sinup page and added login backend
USer can login using the login screen. IF the email is not found, it redirect to the signup page. On the signup page user enter details and a new user is created in the system and user is logged in. The logout url is implemented but the button is still pending.