FEAT: Basic Settings page
On the settings page, the user can view the firstname lastname and edit those. In addition to tat user can view or edit profile photo. The Logic for profile phtoto is still pending. The username and email are also displayed but user can't change those since we will use those to generate user specific file logics.