FEAT: Converted P1 for course builder to forms and added course category addition
The Category addition add the selected categorywhen add button is pressed. The selected cat is displyed in div with other selected categories. A remove button is peresent is eaach cat to remove it from selection. Do duplicate category can be selected. Nor can blank item be added to the div.